+++ server/routes/index.ts
@@ -1,5 +1,7 @@
 import { router as layoutRoutes } from "./layoutRoutes"
 import { router as leadRoutes } from "./leadRoutes"
 import { router as siteRoutes } from "./siteRoutes"
+import { router as tokenRoutes } from "./tokenRoutes"
+import { router as baseRoutes } from "./baseRoutes"
 
-export { layoutRoutes, leadRoutes, siteRoutes }
+export { layoutRoutes, leadRoutes, siteRoutes,tokenRoutes,baseRoutes }
